Get project into working state with embedded figwheel repl
So this is a big commit.  It has in it:

1. Remove lessons that are majorly broken (hopefully can remerge in from upstream)
2. Fix lessons that have minor brokeness
3. Remove the current reloading mechanisms
4. Rewrite the dev/test/prod setup to use user.clj and a component arch
5. Cleanup project.clj into a minimal version
6. Require checkouts for gamma and gamma-driver

The dev environment works mostly well.  It starts in the user namespace
and functions `go` `reset` `browser-repl` and `fig-status` work. Since this
uses a component arch, reloading clj files is done by calling `reset`.

There is one complication with `reset`, in that http-kit does not close websockets
when it is stopped.  This causes figwheel's liveness pings to error. To prevent
this basically requires closing the browser tab.

Testing now works as well.

Unfortunatly production is not tested yet. `lein uberjar` does not allow checkouts,
so this requires waiting until a new gamma and gamma driver are released.